Integration tests: remove driver_install_or_update_darwin_test.go
